+++ tests/test_cases/lrn_gpu_test.cpp
@@ -71,7 +71,7 @@ TEST(local_response_normalization_gpu, lrn_test) {
     // ------------------------------------------------------------------------------------------------
     // test run
     auto outputs = network.execute();
-    EXPECT_EQ(outputs.size(), 1);
+    EXPECT_EQ(outputs.size(), size_t(1));
     EXPECT_EQ(outputs.begin()->first, "lrn");
 
     auto output = outputs.begin()->second.get_memory();
@@ -164,7 +164,7 @@ TEST(local_response_normalization_gpu, lrn_test_batches) {
     // ------------------------------------------------------------------------------------------------
     // test run
     auto outputs = network.execute();
-    EXPECT_EQ(outputs.size(), 1);
+    EXPECT_EQ(outputs.size(), size_t(1));
     EXPECT_EQ(outputs.begin()->first, "lrn");
 
     auto output = outputs.begin()->second.get_memory();
